Factor out ensureOlmSessionsForDevices
... and move to olmlib
